requirements.txt
chkcsv==0.8.0.2
mysql-connector==2.1.6
mysql-connector-python==8.0.12
protobuf==3.6.1
six==1.11.0
